refactor(vim/tmux): use Shift-arrows when need to move betwen split/panes
I tried to stick to Ctrl+hjkl, but with Colemak layout it is not that convenient. Additionally,
there are quite a few default mappings that this shadows, e.g. clear screen, move up within fzf
search list with Ctrl+k. On top of that, I would constantly try to use Shift-Right/Left to switch
betwen tmux panes (since I've been using that for quite sometimes), but in the current setup it
would switch between tmux windows :angry:. In any case, I can try to get used to <C-W>hjkl for
moving between vim splits, since that is a default mapping)